Update Ruby commands in homebrew workflow
The matrix values have been updated from 'rubyfile' to 'command' for increased clarity in the GitHub Actions homebrew workflow. This allows for more flexible use of download and checksum calculation by specifying the command instead of the ruby file. Also, the download URL format has been changed to reflect the correct path for the specific command.